backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
flask==3.1.1
flask-jwt-extended==4.7.1
flask-login==0.6.3
flask-migrate==4.1.0
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.2
mysql-connector-python==9.3.0
pydantic==2.11.5
pydantic-core==2.33.2
pyjwt==2.10.1
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
sqlalchemy==2.0.41
sqlmodel==0.0.24
typing-extensions==4.14.0
typing-inspection==0.4.1
werkzeug==3.1.3
wsproto==1.2.0
